modify toolchain

+++ components/drivers/misc/adc.c
@@ -48,11 +48,11 @@ static rt_err_t _adc_control(rt_device_t dev, int cmd, void *args)
 
     if (cmd == RT_ADC_CMD_ENABLE && adc->ops->enabled)
     {
-        result = adc->ops->enabled(adc, (rt_uint32_t)args, RT_TRUE);
+        result = adc->ops->enabled(adc, (rt_uint32_t)(intptr_t)args, RT_TRUE);
     }
     else if (cmd == RT_ADC_CMD_DISABLE && adc->ops->enabled)
     {
-        result = adc->ops->enabled(adc, (rt_uint32_t)args, RT_FALSE);
+        result = adc->ops->enabled(adc, (rt_uint32_t)(intptr_t)args, RT_FALSE);
     }
     else if (cmd == RT_ADC_CMD_GET_RESOLUTION && adc->ops->get_resolution && args)
     {
